BgDivider && MoreOptionsIcon UI components, useGetCurrentUser custom hook
