+++ src/components/FavouriteArtistCards.tsx
@@ -23,6 +23,7 @@ import NoImage from "../assets/no_artist.jpg";
 
 import 'swiper/css';
 import "swiper/css/effect-cards";
+import { useGetCurrentUser } from "../hooks/useGetCurrentUser";
 interface Props {
     selectedFavouriteArtistInfo: MainDatum | null;
     setSelectedFavouriteArtistInfo: React.Dispatch<React.SetStateAction<MainDatum | null>>
@@ -32,14 +33,7 @@ interface Props {
 export default function FavouriteArtistCards(props: Props) {
     const [favouriteArtists, setFavouriteArtists] = useState<DocumentData[]>([]);
 
-    const { firebaseUsers: users } = useSelector((state: FirebaseUsersSelectorInterface) => state.firebaseUsers);
-
-    const { user: userData } = useSelector((state: UserAuthSelector) => state.userAuth);
-    const user: User = JSON.parse(userData as string);
-
-
-    const [firebaseUser] = useState<UserDoc>(users.find(usr => usr.data.email === user?.email));
-
+    const [user, firebaseUser] = useGetCurrentUser();
 
     const favouriteArtistsCollection = collection(firebaseDatabase, 'users_favourite_artist');
