+++ src/components/UI/BgDivider/BgDivider.tsx
@@ -1,0 +1,5 @@
+export default function BgDivider() {
+    return (
+        <div className="w-full h-28 bg-gradient-to-l from-transparent to-black sm:h-52" />
+    );
+}
